Extract super class of Hiragana and Katakana Modes
